split git clone into a separate module
